requirements.txt
python-telegram-bot==20.7
aiohttp
yt-dlp
redis
python-dotenv